framework: Fix for making SAP sdk code 64 bit compliant
Fix for making SAP SDK code 64 bit compliant. The typedefs
were incorrect on an LP64 architecture since a long int will
be 64-bits.

CRs-Fixed: 640894

+++ softap/sdk/qsap_api.h
@@ -39,8 +39,8 @@ typedef unsigned char      u8;
 typedef char               s8;
 typedef unsigned short int u16;
 typedef signed short int   s16;
-typedef unsigned long int  u32;
-typedef signed long int    s32;
+typedef unsigned int       u32;
+typedef signed int         s32;
 
 /** Success and error messages */
 #define SUCCESS                     "success"
